requirements.txt
anthropic>=0.40.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
plotly>=5.17.0
python-dotenv>=1.0.0
